Refactor productConfig form into separate component for use outside the productConfigModal
