+++ src/app/checkout/checkout.tpl.html
@@ -24,11 +24,11 @@
           <div class="col-md-12">
             <div class="panel">
               <div class="panel-body">
-                <h3 class="panel-name">Checkout</h3>
+                <h3 class="panel-name" translate>Checkout</h3>
 
                 <div class="steps-wrap mb_x clearfix">
                   <div class="steps-single" ng-class="{'on': $ctrl.checkoutStep === 'contact'}">
-                    <span class="step-title">
+                    <span class="step-title" translate>
                     Contact Information
                     </span>
                     <div class="step-bar">
@@ -36,7 +36,7 @@
                     </div>
                   </div>
                   <div class="steps-single" ng-class="{'on': $ctrl.checkoutStep === 'payment'}">
-                    <span class="step-title">
+                    <span class="step-title" translate>
                     Payment Method
                     </span>
                     <div class="step-bar">
@@ -44,7 +44,7 @@
                     </div>
                   </div>
                   <div class="steps-single" ng-class="{'on': $ctrl.checkoutStep === 'review'}">
-                    <span class="step-title">
+                    <span class="step-title" translate>
                     Review and Submit Gift
                     </span>
                     <div class="step-bar">
