+++ src/app/productConfig/giveGiftParams.js
@@ -1,0 +1,8 @@
+export const giveGiftParams = {
+  designation:  'd',
+  campaignPage: 'c',
+  amount:       '$',
+  frequency:    'f',
+  day:          'dd',
+  campaignCode: 'CampaignCode'
+};
