+++ src/app/productConfig/productConfig.component.js
@@ -4,7 +4,7 @@ import template from './productConfig.tpl.html';
 import commonModule from 'common/common.module';
 import productModalService from 'common/services/productModal.service';
 import modalStateService from 'common/services/modalState.service';
-import {giveGiftParams} from './productConfigModal/productConfig.modal.component';
+import {giveGiftParams} from './giveGiftParams';
 import analyticsFactory from 'app/analytics/analytics.factory';
 
 //include designation edit button component to be included on designation page
@@ -22,37 +22,12 @@ class ProductConfigController {
     this.$log = $log;
   }
 
-  $onInit() {
-    this.loadingModal = false;
-  }
-
   configModal() {
-    this.loadingModal = true;
-    this.error = false;
-    let modalInstance = this.productModalService
+    this.productModalService
       .configureProduct( this.productCode, {
         'campaign-code': this.campaignCode,
         'campaign-page': this.campaignPage
       }, false );
-    if(!modalInstance){
-      // Another modal already opened
-      this.loadingModal = false;
-      return;
-    }
-    modalInstance.rendered.then( () => {
-      this.loadingModal = false;
-      this.analyticsFactory.giveGiftModal(this.productCode);
-    }, angular.noop );
-    modalInstance.result.then( () => {
-        this.$window.location = '/cart.html';
-      },
-      reason => {
-        if(reason && reason !== 'escape key press' && reason !== 'backdrop click'){ // Avoid labeling normal closes, escape key, or backdrop clicks as errors
-          this.$log.error('Error opening product config modal', reason);
-          this.error = true;
-        }
-        this.loadingModal = false;
-      });
   }
 }
