+++ src/common/services/fixtures/designation.infinity.fixture.js
@@ -1,0 +1,35 @@
+export default {
+  "jcr:primaryType": "cq:Page",
+  "jcr:createdBy":   "admin",
+  "jcr:created":     "Thu Nov 10 2016 17:03:19 GMT+0000",
+  "jcr:content":     {
+    "jcr:primaryType":         "cq:PageContent",
+    "jcr:mixinTypes":          ["mix:versionable"],
+    "startDate":               "2016-07-01T00:00:00.000+00:00",
+    "jcr:createdBy":           "admin",
+    "jcr:title":               "PowerPacksTM for Inner City Children",
+    "customized":              true,
+    "jcr:versionHistory":      "671c8c7b-74af-47d4-8f6d-3a53e103ae3e",
+    "cq:template":             "/apps/Give/templates/campaign",
+    "secondaryPhoto":          "/content/dam/give/designations/0/1/2/3/4/0123456/CMS1_085319.jpg",
+    "psText":                  "",
+    "organizationId":          "1-103-1",
+    "parentDesignationNumber": "",
+    "paragraphText":           "<p>Sample Text</p>",
+    "jcr:predecessors":        ["df5a7547-4d83-451b-bbc0-179362278d75"],
+    "jcr:created":             "Mon Nov 21 2016 21:46:54 GMT+0000",
+    "designationNumber":       "0123456",
+    "designationName":         "PowerPacksTM for Inner City Children",
+    "designationType":         "Campaign",
+    "jcr:baseVersion":         "df5a7547-4d83-451b-bbc0-179362278d75",
+    "jcr:isCheckedOut":        true,
+    "vanityURL":               "0123456_9876",
+    "hideInNav":               true,
+    "jcr:uuid":                "11221ed0-a53b-4a10-8ef3-bc13908fc241",
+    "sling:resourceType":      "Give/components/page/campaign",
+    "letterDate":              "",
+    "websiteURL":              "",
+    "defaultCampaign":         "867EM1",
+    "suggestedAmounts":        {"jcr:primaryType":"nt:unstructured"}
+  }
+};
